Implementing review suggestions
