+++ vantage6-algorithm-store/vantage6/algorithm/store/model/__init__.py
@@ -1,5 +1,6 @@
 # pylint: disable=unused-import
 from vantage6.backend.common.base import Base
+from vantage6.backend.common.permission_models import role_rule_association
 from vantage6.algorithm.store.model.algorithm import Algorithm
 from vantage6.algorithm.store.model.argument import Argument
 from vantage6.algorithm.store.model.database import Database
@@ -12,4 +13,3 @@ from vantage6.algorithm.store.model.user import User
 from vantage6.algorithm.store.model.review import Review
 from vantage6.algorithm.store.model.ui_visualization import UIVisualization
 from vantage6.algorithm.store.model.policy import Policy
-from vantage6.backend.common.permission_models import role_rule_association
