+++ vantage6-algorithm-store/vantage6/algorithm/store/resource/algorithm.py
@@ -25,9 +25,7 @@ from vantage6.algorithm.store.resource import (
 )
 
 # TODO move to common / refactor
-from vantage6.algorithm.store.resource import (
-    AlgorithmStoreResources,
-)
+from vantage6.algorithm.store.resource import AlgorithmStoreResources
 from vantage6.algorithm.store.permission import (
     PermissionManager,
     Operation as P,
