+++ vantage6-server/vantage6/server/__init__.py
@@ -9,10 +9,6 @@ authenticated nodes and users via the socketIO server that is run here.
 import os
 from gevent import monkey
 
-from vantage6.server.model import Role, Rule
-from vantage6.server.model.rule import Operation, Scope
-from vantage6.server.permission import PermissionManager
-
 # This is a workaround for readthedocs
 if not os.environ.get("READTHEDOCS"):
     # flake8: noqa: E402 (ignore import error)
@@ -52,18 +48,22 @@ from pathlib import Path
 from sqlalchemy.orm.exc import NoResultFound
 
 from vantage6.common import logger_name
-from vantage6.common.serialization import jsonable
 from vantage6.common.globals import PING_INTERVAL_SECONDS
 from vantage6.backend.common.globals import HOST_URI_ENV
-from vantage6.server import db
-from vantage6.cli.context.server import ServerContext
+from vantage6.backend.common.jsonable import jsonable
 from vantage6.backend.common.base import DatabaseSessionManager, Database
 from vantage6.backend.common.permission import RuleNeed
+from vantage6.server.model import Role, Rule
+from vantage6.server.model.rule import Operation, Scope
+from vantage6.server.permission import PermissionManager
+from vantage6.server import db
+from vantage6.cli.context.server import ServerContext
 from vantage6.server.resource.common.output_schema import HATEOASModelSchema
 from vantage6.server.globals import (
     APPNAME,
     ACCESS_TOKEN_EXPIRES_HOURS,
     RESOURCES,
+    RESOURCES_PATH,
     SUPER_USER_INFO,
     REFRESH_TOKENS_EXPIRE_HOURS,
     DEFAULT_SUPPORT_EMAIL_ADDRESS,
@@ -138,9 +138,7 @@ class ServerApp:
         self.socketio = self.setup_socket_connection()
 
         # setup the permission manager for the API endpoints
-        self.permissions = PermissionManager(
-            "vantage6.server.resource", RESOURCES, DefaultRole
-        )
+        self.permissions = PermissionManager(RESOURCES_PATH, RESOURCES, DefaultRole)
 
         # Api - REST JSON-rpc
         self.api = Api(self.app)
