+++ vantage6-server/vantage6/server/controller/fixture.py
@@ -3,7 +3,7 @@ import logging
 
 import vantage6.server.model as db
 from vantage6.backend.common.base import Database
-from vantage6.server import RESOURCES, DefaultRole
+from vantage6.server import RESOURCES, RESOURCES_PATH, DefaultRole
 from vantage6.server.permission import PermissionManager
 from vantage6.common.task_status import TaskStatus
 from vantage6.common.serialization import serialize
@@ -55,9 +55,9 @@ def load(fixtures: dict, drop_all: bool = False) -> None:
     # permissions.
     if not db.Rule.get():
         permissions = PermissionManager(
-            "vantage6.server.resource", RESOURCES, DefaultRole
+            RESOURCES_PATH, RESOURCES, DefaultRole
         )
-        permissions.load_rules_from_resources("vantage6.server.resource", RESOURCES)
+        permissions.load_rules_from_resources(RESOURCES_PATH, RESOURCES)
 
     log.info("Create Organizations and Users")
     for org in fixtures.get("organizations", {}):
